fixes 2

+++ Content/Buffs/PetTurkey.cs
@@ -9,14 +9,15 @@ namespace Consolaria.Content.Buffs
 		public override void SetStaticDefaults() {
 			DisplayName.SetDefault("Pet Turkey");
 			Description.SetDefault("Big and meaty turkey");
+
 			Main.buffNoTimeDisplay[Type] = true;
 			Main.vanityPet[Type] = true;
 		}
 
 		public override void Update(Player player, ref int buffIndex){ 
 			player.buffTime[buffIndex] = 18000;
-			sbyte type = (sbyte)ModContent.ProjectileType<Projectiles.Friendly.Pets.PetTurkey>();
-			//var entitySource = player.GetSource_Buff(buffIndex);
+			ushort type = (ushort)ModContent.ProjectileType<Projectiles.Friendly.Pets.PetTurkey>();
+
 			if (player.whoAmI == Main.myPlayer && player.ownedProjectileCounts[type] <= 0)
 				Projectile.NewProjectile(player.GetSource_Buff(buffIndex), player.Center, Vector2.Zero, type, 0, 0f, player.whoAmI);
 		}
